+++ R/solveThreeStateModel.R
@@ -1,18 +1,18 @@
-#' Solve Parameters in a Three-State Ill-Death Model
+#' Solve Parameters in a Three-State Illness-death Model
 #'
 #' @description
-#' The ill-death model consists of three states, \code{initial}, \code{progression},
+#' The illness-death model consists of three states, \code{initial}, \code{progression},
 #' and \code{death}. It can be used to model the progression-free survival (PFS)
 #' and overall survival (OS) in clinical trial simulation. It models the
-#' correlation PFS and OS without assumptions on latent status and copula.
+#' correlation between PFS and OS without assumptions on latent status and copula.
 #' Also, it does not assume PFS and OS satisfy the proportional hazard assumption
-#' simultaneously. The three-state ill-death model ensures a nice property that
+#' simultaneously. The three-state illness-death model ensures a nice property that
 #' PFS <= OS with probability one. However, it requires three hazard parameters
 #' under the homogeneous Markov assumption. In practice, hazard parameters are
 #' hard to specify intuitively especially when no trial data is available at
 #' the planning stage.
 #'
-#' This function reparametrizes the ill-death model in term of three parameters,
+#' This function reparametrizes the illness-death model in term of three parameters,
 #' i.e. median of PFS, median of OS, and correlation between PFS and OS. The
 #' output of this function, which consists of the three hazard parameters, can
 #' be used to generate PFS and OS with desired property. It can be used with
@@ -20,7 +20,7 @@
 #' endpoints in \code{TrialSimulator}.
 #'
 #' For more information, refer to
-#' \href{https://zhangh12.github.io/TrialSimulator/articles/simulatePfsAndOs.html}{this vignette}.
+#' \href{https://zhangh12.github.io/TrialSimulator/articles/simulatePfsAndOsIdm.html}{this vignette}.
 #'
 #' @param median_pfs numeric. Median of PFS.
 #' @param median_os numeric. Median of OS.
@@ -126,7 +126,7 @@ solveThreeStateModel <- function(median_pfs, median_os, corr,
 
 }
 
-#' Plot result of three-state ill-death model
+#' Plot result of three-state illness-death model
 #'
 #' @param x an object returned by \code{solveThreeStateModel()}.
 #' @param ... currently not supported.